Refactoring code
- added config file
- added dynamodb file for making calls
- added utils.py
