+++ src/app/config.py
@@ -1,0 +1,16 @@
+import os
+
+# vars
+DISCORD_PUBLIC_KEY = os.environ.get("DISCORD_PUBLIC_KEY")
+GOOGLE_BOOKS_API_URL = "https://www.googleapis.com/books/v1/volumes"
+GREETINGS = [
+    "Hello there!",
+    "Hi!",
+    "Good to see you!",
+    "Greetings!",
+    "Hey, how are you?",
+    "Nice to meet you!",
+    "Hello, friend!",
+    "Hi, hope you’re doing well!",
+]
+EMOJIS = ["👋", "😊", "🙌", "🌟", "🤗", "😄", "✨", "😎", "😁"]
