+++ src/app/dynamodb.py
@@ -1,0 +1,32 @@
+import boto3
+from datetime import datetime, timezone
+
+dynamodb = boto3.resource("dynamodb")
+# @TODO: Change this after you have tested to accomodate for prod
+book_table = dynamodb.Table("Alpha-BookClubHistory")
+
+
+def put_book(guild_id, user_id, selected_book, discussion_date):
+    # Extract info from selected_book (adjust fields as per your data)
+    volume_info = selected_book.get("volumeInfo", {})
+    title = volume_info.get("title", "Unknown Title")
+    authors = ", ".join(volume_info.get("authors", ["Unknown Author"]))
+    isbn_list = volume_info.get("industryIdentifiers", [])
+    isbn_13 = next((id["identifier"] for id in isbn_list if id["type"] == "ISBN_13"), None)
+    isbn_10 = next((id["identifier"] for id in isbn_list if id["type"] == "ISBN_10"), None)
+    isbn = isbn_13 or isbn_10 or "N/A"
+
+    # Create a unique primary key for current selection (e.g., "CURRENT_BOOK")
+
+    # Put item into DynamoDB table
+    book_table.put_item(
+        Item={
+            "guild_id": guild_id,  # Partition key (table schema dependent)
+            "discussion_date": discussion_date,
+            "title": title,
+            "authors": authors,
+            "isbn": isbn,
+            "set_by_user": user_id,
+            "timestamp": datetime.now(timezone.utc).isoformat(),
+        }
+    )
